Change how we determine to extract the file or not
This adds a new property :overwrite which controls the destination overwrite behavior. If set to false we will never overwrite. If set to true we will always overwrite. If set to :auto we'll use the old behavior of checking the mtime of the archive vs. the disk. The added benefit here is that with this refactor you don't have to do that time comparison if there's nothing on disk. Seemed kind of silly to do that when were guaranteed to extract.

Signed-off-by: Tim Smith <tsmith@chef.io>